The jar of the app
Without lesson from teachers, it was really hard to create a jar which can read and write in file which is in the jar's file. With the lack of time, the selected solution has been to create a data file outside of the jar's file, this is not elegant but it works.